support CCSequence & CCAnimation create with Vector
fixed Flog main target being deleted before use.

fixed effect frame count.

+++ projects/NarutoSenki/Classes/Effect.h
@@ -11,7 +11,7 @@ public:
 	static Effect *create(const string &szImage, CharacterBase *attacker);
 
 private:
-	FiniteTimeAction *createEffectAnimation(const string &file, int frameCount, float fps, bool isRepeat);
+	FiniteTimeAction *createEffectAnimation(const string &file, size_t frameCount, float fps, bool isRepeat);
 	FiniteTimeAction *createFontAnimation();
 	void removeEffect();
 	void removeFontEffect();
